jooq plus fixes

+++ project/scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/dto/domain/Link.java
@@ -3,6 +3,7 @@ package ru.tinkoff.edu.java.scrapper.dto.domain;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
+import org.jooq.impl.QOM;
 
 import java.net.URI;
 import java.time.OffsetDateTime;
@@ -14,6 +15,6 @@ public class Link {
     private long id;
     private URI url;
     private OffsetDateTime lastUpdate;
-    private long answerCount;
-    private long commentCount;
+    private Integer answerCount;
+    private Integer commentCount;
 }
